refactor palette panel into modular components
